toggle functionality on displayed
