+++ public/index.html
@@ -42,7 +42,7 @@
           <button type="submit">Submit</button>
         </form>
       </div>
-      <div id="books-list"></div>
+      <div id="books-list" style="display: none"></div>
     </section>
     <section id="participants">
       <h2>Participants</h2>
@@ -59,14 +59,14 @@
           <button type="submit">Submit</button>
         </form>
       </div>
-      <div id="participants-list"></div>
+      <div id="participants-list" style="display: none"></div>
     </section>
 
     <section id="roles">
       <h2>Roles</h2>
       <button id="loadRoles">Load Roles</button>
       <button id="addRoleBtn">Add Role</button>
-      <ul id="roles-list"></ul> 
+      <ul id="roles-list" style="display: none"></ul> 
       <div id="role-form-container" style="display: none">
         <form id="role-form">
           <input
@@ -78,7 +78,6 @@
           <button type="submit">Submit</button>
         </form>
       </div>
-      <div id="roles-list"></div>
       <div id="role-details"></div>
     </section>
