+++ src/utils/helpers/roleHelpers.js
@@ -1,39 +1,4 @@
-export function toggleForm(containerId) {
-    const container = document.getElementById(containerId);
-    if (!container) {
-        console.error(`toggleForm: No element found with ID '${containerId}'`);
-        return;
-    }
-    container.style.display = container.style.display === 'none' ? 'block' : 'none';
-}
-
-export function displayBooks(books) {
-    const list = document.getElementById('books-list');
-    if (!list) {
-        console.error('displayBooks: No element found with ID "books-list"');
-        return;
-    }
-    list.innerHTML = '';
-    books.forEach(book => {
-        const item = document.createElement('li');
-        item.textContent = `${book.title} - ISBN: ${book.isbn}`;
-        list.appendChild(item);
-    });
-}
-
-export function displayParticipants(participants) {
-    const list = document.getElementById('participants-list');
-    if (!list) {
-        console.error('displayParticipants: No element found with ID "participants-list"');
-        return;
-    }
-    list.innerHTML = '';
-    participants.forEach(participant => {
-        const item = document.createElement('li');
-        item.textContent = participant.name;
-        list.appendChild(item);
-    });
-}
+import { displayStatusMessage } from "./common.js";
 
 export function displayRoles(roles) {
     const list = document.getElementById('roles-list');
@@ -103,36 +68,4 @@ export function deleteRole(role) {
             console.error('Error deleting role:', error);
             displayStatusMessage('Error deleting role: ' + error.message, 'error');
         });
-}
-
-export function displayStatusMessage(message, status) {
-    const statusDiv = document.getElementById('status-messages');
-    statusDiv.textContent = message;
-    statusDiv.className = status; // Apply different classes for styling based on success or error
-}
-
-
-export function displayBookParticipants(bookParticipants, bookId) {
-    const list = document.getElementById('book-participants-list');
-    if (!list) {
-        console.error('displayBookParticipants: No element found with ID "book-participants-list"');
-        return;
-    }
-    list.innerHTML = '';
-    if (!bookParticipants || bookParticipants.length === 0) {
-        const item = document.createElement('li');
-        item.textContent = 'No participants found for this book.';
-        list.appendChild(item);
-        return;
-    }
-    bookParticipants.forEach(bp => {
-        const item = createBookParticipantItem(bp, bookId);
-        list.appendChild(item);
-    });
-}
-
-function createBookParticipantItem(bp, bookId) {
-    const item = document.createElement('li');
-    item.textContent = `Book ID: ${bookId}, Participant: ${bp.participant.name} (ID: ${bp.participant.participantid}), Role: ${bp.role.description} (ID: ${bp.role.roleid})`;
-    return item;
 }
